Organize theme styles into separate files
