+++ services/ui-src/src/styles/foundations/colors.ts
@@ -1,0 +1,59 @@
+export const colors = {
+  palette: {
+    // primary
+    primary: "#0071bc",
+    primary_darker: "#004f84",
+    primary_darkest: "#00395e",
+    // secondary
+    secondary_lightest: "#e6f9fd",
+    secondary_lighter: "#b3ecf8",
+    secondary_light: "#4ed2ee",
+    secondary: "#02bfe7",
+    secondary_dark: "#02acd0",
+    secondary_darker: "#0186a2",
+    secondary_darkest: "#016074",
+    // status: success
+    success_lightest: "#e7f3e7",
+    success_lighter: "#89c487",
+    success_light: "#2a9526",
+    success: "#12890e",
+    success_dark: "#107b0d",
+    success_darker: "#0d600a",
+    success_darkest: "#094507",
+    // status: warn
+    warn_lightest: "#fef9e9",
+    warn_lighter: "#fce28f",
+    warn_light: "#f9ca35",
+    warn: "#f8c41f",
+    warn_dark: "#dfb01c",
+    warn_darker: "#ae8916",
+    warn_darkest: "#7c6210",
+    // status: error
+    error_lightest: "#fce8ec",
+    error_lighter: "#f7bbc5",
+    error_light: "#f18e9e",
+    error: "#e31c3d",
+    error_dark: "#cc1937",
+    error_darker: "#9f142b",
+    error_darkest: "#720e1f",
+    // neutrals
+    white: "#ffffff",
+    gray_lightest: "#f2f2f2",
+    gray_lighter: "#d9d9d9",
+    gray_light: "#a6a6a6",
+    gray: "#5a5a5a",
+    gray_dark: "#404040",
+    base: "#262626",
+    black: "#000000",
+    // other
+    focus_light: "#ffffff",
+    focus_dark: "#bd13b8",
+    muted: "#e9ecf1",
+    // custom
+    gray_lightest_highlight: "#f8f8f8",
+    gray_medium: "#71767a",
+    gray_medium_dark: "#5B616B",
+    spreadsheet: "#1d6f42",
+    spreadsheet_dark: "#174320",
+  },
+};
