+++ services/ui-src/src/styles/foundations/index.ts
@@ -1,0 +1,9 @@
+import { breakpoints } from "./breakpoints";
+import { fonts } from "./fonts";
+import { lineHeights } from "./lineHeights";
+
+export const foundations = {
+  breakpoints: breakpoints,
+  fonts: fonts,
+  lineHeights: lineHeights,
+};
