+++ services/ui-src/src/styles/foundations/lineHeights.ts
@@ -1,0 +1,8 @@
+export const lineHeights = {
+  // https://design.cms.gov/utilities/line-height/
+  normal: "normal",
+  reset: 1,
+  heading: 1.3,
+  base: 1.5,
+  lead: 1.7,
+};
